Split the test into two so ctest can run them in parallel.
